.github/skills/standard/scripts/requirements.txt
azure-identity>=1.21.0
python-dotenv>=1.1.0
requests>=2.32.3